fix(symbolic): give Bool leaves their own Z3 constant (angr-9ke6b.223)
RustBV has no Bool sort — a claripy Bool leaf is modelled as a width-1
Symbolic — and RustBV::from_parts builds a leaf's Z3 term from its NAME.
So BVS("x", 1, explicit_name=True) and BoolS("x", explicit_name=True),
already given separate rust_ids by the angr-9ke6b.38 registry sort tag,
still collapsed to one BV::new_const("x", 1): constraining the BV to 1
forced the Bool true.

The name now carries the sort. SymbolKind::rust_symbol_name tags a Bool
leaf's Rust-side name and strip_bool_symbol_name inverts it;
import_symbolic_leaf applies the tag before any registry call, so the
name_to_info key, rust_id_to_name, the RustBV name and the Z3 term all
agree (the precondition the by-hash rebind in angr-izov2 relies on), and
the tag survives a RustBVData snapshot round-trip for free since the name
is the one field every from_parts caller already carries. BitVector is the
identity, so no existing name or Z3 term changes.

from_parts decodes the tag to Bool::new_const(claripy_name).ite(1, 0).
Using claripy's own name for the Bool constant is load-bearing, not
cosmetic: claripy's z3 backend encodes BoolS("x") as a Bool-sorted const
named "x" (verified), and RustSolverContext::add_constraint_ast has a
raw-passthrough fast path that hands that term straight to the solver.
Building BV::new_const("x", 1) instead — as before this commit — named an
unrelated declaration, so a Bool constraint added through the fast path
never bound to a leaf imported through claripy_to_rustbv. Both halves are
fixed by the same encoding.

Also drops a redundant u32 cast in subcall_tests.rs that reddened clippy.

Tests: 5 Rust unit tests over the tag round-trip, 2 Python tests over the
end-to-end contract — BVS("x",1)==1 with Not(BoolS("x")) is now SAT, and a
constraint on a bare BoolS leaf binds on re-import (it did not before).

+++ native/angr/src/exploration/subcall_tests.rs
@@ -335,7 +335,7 @@ fn link_register_abi_subcall_roundtrip_resumes_at_lr() {
                 &mut state,
                 NativeSubcall {
                     proc_name: "subcall_test".into(),
-                    saved_args: vec![RustBV::concrete(41, (cc_ptr * 8) as u32)],
+                    saved_args: vec![RustBV::concrete(41, cc_ptr * 8)],
                     caller_return_addr: captured,
                     target: SubcallTestProc::GUEST_TARGET,
                     sub_args: vec![],
